excel_sort_bot/requirements.txt
openpyxl>=3.1.0
python-dateutil>=2.8.0
pytest>=7.0.0

# API dependencies
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# PDF export
fpdf2>=2.7.0

# AI dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# Database (producción)
supabase>=2.0.0
